[buteo-sync-plugin-caldav] Prefix UID of received events with NBUID. Contributes to MER#1465
mkcal doesn't allow events to exist in multiple notebooks.
If the remote server has an event which exists in multiple collections,
downsync will fail due to uniqueness constraint violation in mkcal.

The best solution would be to fix mkcal but that is out of scope.
This commit instead prefixes the event uid with the notebook uid
when downsynced, and extracts only the remote event uid for upsync,
to ensure that the uniqueness constraint is not violated locally.

Note, however, that this means that there will be two (or more)
individual event instances locally, which refer to the same event
remotely, and so this may result in consistency issues (although
they should eventually resolve after two sync cycles).

Contributes to MER#1465