[buteo-sync-plugin-caldav] Improve handling of multi-day all-day events. Contributes to MER#1483
Previously, we needed to set a time into dtStart/dtEnd which were
originally date-only datetimes, because mkcal had a bug which meant
that FloatingDate datetimes were not returned as date-only datetimes
after the first write.

However, that bug has now been fixed in mkcal.
So, this commit ensures that we treat date-only dates as date-only
dates correctly.

Contributes to MER#1483, MER#1492