[buteo-sync-plugin-caldav] Only allow sync if the account is enabled. Contributes to JB#18634
This commit checks to see if the account is enabled, and only performs
sync if it is, otherwise it returns an error.

This is not a problem in the common case (i.e., a scheduled sync) as
Buteo checks before triggering the profile, but in some cases (e.g.,
manually triggered syncs) this check is required.

Contributes to JB#18634