[buteo-sync-plugin-caldav] Handle known-orphan occurrences. Contributes to JB#38601
Previously, we would still attempt to dissociate known-orphan
recurrences from non-existent base recurring series on update.
This commit ensures that if the occurrence is a known orphan,
in the case where no parent base recurring series is found in
the calendar (e.g., it has not been saved as part of a previous
sync either), generate a parent base series with the orphan's
recurrenceId as an RDATE.

This commit also ensures that if a non-critical error occurs
when saving an event (e.g., one specific event cannot be saved)
that we nonetheless continue to process all further resources
instead of immediately stopping.

Contributes to JB#38601