Fix unit test regression
When determining the order in which to handle resources received
from the server, we should (where possible) handle the resources
received first before those received last.

This commit achieves this by iterating in reverse order, so that
as base events are prepend()ed, "earlier" events are prepended
after "later" events.

A more correct way to achieve this would be to ensure that we
order resources based on the SEQUENCE (revision()) field, however
this requires far more computation, since each resource in the
list can contain multiple incidences, and each incidence can have
a different SEQUENCE value.