[buteo-sync-plugin-caldav] Handle broken upsync gracefully. Contributes to JB#38601
Previously, some broken data (e.g., exception occurrences with no
parent series which is downsynced from remote) could cause issues
during upsync (when we try to generate an ICS of the entire series
which contains the modified exception occurrence), possibly resulting
in a crash.

This commit ensures that we handle those cases properly, by ensuring
that we use the base series UID as the instance UID when generating
the series .ics to upload, and also adds some better error handling
in case similar issues are encountered, and adds more debug logging.

Contributes to JB#38601