[buteo-sync-plugin-caldav] Allow sync to continue after broken local modification. Contributes to MER#1944
If a local event has no remote URL stored for it, we cannot upsync
changes for that event to the remote server.  Previously, this case
was handled as an error; this commit ensures that we merely ignore
that change and allow sync to continue, to avoid being "stuck"
forever, even though it is an internal error.