[buteo-sync-plugin-caldav] Send exceptions with base incidences when only the base has changed.
Previously, if the base incidence of a recurring event was modified
(but not any instance of its exceptions), the ICS data were
constructed using the base incidence only, and not the base and the
exception incidences. This is resulting in pushing to the server a
recurring event with without exception incidences, and previous
exception incidences were replaced by nothing at the exception
dates.