[buteo-sync-plugin-caldav] Add CalDAV invitation plugin for mkcal. Contributes to JB#46575
This plugin returns the email address associated with the owner
of the notebook (in most cases, will be the same as the email address
of the account owner, as recorded by the server) for use as the
ORGANIZER field of invitations.

Aside from that, it currently does nothing, as most CalDAV servers
support performing iTIP and/or iMIP scheduling based on changes to
.ics organizer data in user calendars (i.e. clients merely need to
make changes to the event, and then sync the event, e.g. setting
participation status value to respond to an invitation).

In the future, we will extend this plugin to manually perform
SCHEDULE iTIP requests to the appropriate server's INBOX or OUTBOX
collections, or send iMIP requests via QMF (requires linking a
valid email account with the CalDAV account on the device).