[buteo-sync-plugin-caldav] Read user principal mailto address. Contributes to JB#46575
For each calendar we are syncing, determine the user principal mailto
address (i.e. email address) associated with it.  In most cases,
this will be the mailto address of the user principal of the account,
but potentially in some cases (e.g. shared calendars) it may be a
different one.

This will allow us to programmatically determine what the appropriate
"organizer email" is associated with each calendar (notebook).

This commit also fixes parsing of some XML in cases where some
HTTP 404 NOT FOUND status propstat sections are returned by the
server.