[buteo-sync-plugin-caldav] Store remote path at notebook creation. Contributes to TJC#224936
For new notebook, the notebook is added in the applyRemoteChanges()
call. If populating the calendar is failing, the remote path is not
added to the notebook, and at next sync, this notebook will not be
detected as associated to the calendar, and the sync process will
create a new one.