[buteo-sync-plugin-caldav] Remove obsolete tests for IncidenceHandler.
These tests were used when the IncidenceHandler was
providing a copy routine and a comparison routine.
It's not the case anymore and the tests were now
using direct copy and comparison from KCalendarCore.
There is no use to test such operators here.

Moreover, the sync process is not using anymore
any comparison between incidences. It was used
when mKCal was reporting local changes on incidences
when the etag was set. It's not the case anymore.