[buteo-sync-plugin-caldav] Fix upsync of imported exceptions.
The sync code relies on the fact that added
exceptions on device inherit the URI and ETAG
of the parent, when the dissociation code is
called.

This assumption is not true when importing
an exception, like a cancel event received
by email for instance.

This commit ensures that such an imported
event is actually seen as a local addition
and that the ETAG used to push the event to
the server is using the ETAG of the parent.