[buteo-sync-plugin-caldav] Correct URL in case of mangled UIDs.
When receiving a new incidence from upstream,
the UID is mangled with the notebook UID to avoid
colision with an existing UID in another notebook.

The URL of the received incidence is stored with
the incidence, so when modifying such incidence
the update is sent to the correct URL and not
one created from the mangled UID.

Unfortunately, creating an exception for such a
recurring incidence on device, will not copy
the URL (as it is stored in comments) and when
pushing the exception to the server, the mangled
UID is used to forge the URL, which is faulty.

See calendarworker.cpp in nemo-qml-plugin-calendar,
the dissociateSingleOccurrence() call is properly
copying the comments, but the information is not
passed to the manager via CalendarData::Event
structure. Then, when saving the event in saveEvent(),
the proper URL is lost.