[buteo-sync-plugin-caldav] Retrieve allowed components per calendar.
Use propfind method 'supported-calendar-component-set'
to inquire the allowed components, like only VTODOS
for task calendars for instance.

These properties are not used to filter out
downsynced or upsynced components yet. The
information is stored to mKCal though and
may be used at UI level to hide calendars
not suited for VEVENT storage.