[buteo-sync-plugin-caldav] Add a specific case for memotoo.com.
Memotoo.com has only one calendar, but it is
listing all the event categories as calendars.
While this could be used to view events with
the right color, as belonging to different
calendars, it is creating an issue when sending
new events to the server : they must contain the
right category in the ICS data.

The easiest way to avoid dealing with categories
is to list only one calendar. Which is what the
service provide. And don't falsely report
categories as calendar.

So filter the calendar list response in the memotoo
case.